Fixed broken tests (#24)
Tests were failing due to a a cleanup error. As the test script iterated through the definitions to install/generate, some of the following definitions would reference previously iterated definitions; those of which helm_repos and components directories were not deleted. This would cause git errors as it would attempt to clone the repository again into a directory already containing a git repo.